pyproject.toml
[project]
name = "mohirdev-culon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-jazzmin>=3.0.4",
    "django-modeltranslation>=0.20.2",
    "django-redis>=6.0.0",
    "django-rosetta>=0.10.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.8",
]
